+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Add change organization status metrics(approved or declined)
+
 ## [0.39.0] - 2023-09-25
 
 ### Added